requirements.txt
mysql-connector-python==8.0.30
psycopg2-binary==2.9.3
pandas==1.5.0
